requirements.txt
torchdata==0.7.0
pydantic==2.9.2
rdkit==2024.3.5
dgllife==0.3.2
ogb==1.3.6
pandas==2.2.3
tensorboard==2.17.1
matplotlib==3.9.2
chardet==5.2.0
trzsz==1.1.5
nvitop==1.3.2
numpy==1.26.4
-e .